VERSION 1.0
Initial release. Beware of the possible fps drop caused by the waypoints of the IDF patrols along the border. Some air to air conflict may occurs due to the close range of the patrols near the border.
There are various skins from the siryan and idf factions that can be downloaded from the DCS user file section of DCS WORLD main website.